Rename "Form Group" to "Form and Formset", and "Form" to "Form Section"
Form Group is named formset so that the association between forms and
sections -- ie. the sections on a form -- is cleaner (FormsetSection as
opposed to FormSections). Otherwise there would be virtually no discernable
difference between FormSection and FormSections, which are functionally
completely different. Therefore, the latter is FormsetSections.
