+++ include/staff/dynamic-form-sections.inc.php
@@ -1,15 +1,15 @@
 <div style="width:700;padding-top:5px; float:left;">
- <h2>Form Groups</h2>
+ <h2>Dynamic Form Sections</h2>
 </div>
 <div style="float:right;text-align:right;padding-top:5px;padding-right:5px;">
- <b><a href="dynamic-forms.php?a=add" class="Icon">Add Dynamic Form Group</a></b></div>
+ <b><a href="dynamic-form-sections.php?a=add" class="Icon">Add New Dynamic Form Section</a></b></div>
 <div class="clear"></div>
 
 <?php
 $page = 1;
 $pagesize = 25;
 $start = ($page-1)*$pagesize;
-$count = DynamicFormGroup::count();
+$count = DynamicFormSection::count();
 $end = min($start+$pagesize, $count);
 $showing = sprintf("Showing %s - %s of %s", $start+1, $end, $count);
 ?>
@@ -24,7 +24,7 @@ $showing = sprintf("Showing %s - %s of %s", $start+1, $end, $count);
         </tr>
     </thead>
     <tbody>
-    <?php foreach (DynamicFormGroup::all() as $form) { ?>
+    <?php foreach (DynamicFormSection::all() as $form) { ?>
         <tr>
             <td/>
             <td><a href="?id=<?php echo $form->get('id'); ?>"><?php echo $form->get('title'); ?></a></td>
