+++ include/staff/helptopic.inc.php
@@ -205,18 +205,18 @@ $info=Format::htmlchars(($errors && $_POST)?$_POST:$info);
         </tr>
        <tr>
            <td><strong>Formset</strong>:</td>
-           <td><select name="form-group-id">
+           <td><select name="formset-id">
                <option value="0">&mdash; Select a Formset &mdash;</option>
-               <?php foreach (DynamicFormGroup::all() as $group) { ?>
+               <?php foreach (DynamicFormset::all() as $group) { ?>
                    <option value="<?php echo $group->get('id'); ?>"
-                       <?php if ($group->get('id') == $info['form_group_id'])
+                       <?php if ($group->get('id') == $info['formset_id'])
                             echo 'selected="selected"'; ?>>
                        <?php echo $group->get('title'); ?>
                    </option>
                <?php } ?>
                </select>
                <em>Information for tickets associated with this help topic</em>
-               &nbsp;<span class="error">&nbsp;<?php echo $errors['form-group-id']; ?></span>
+               &nbsp;<span class="error">&nbsp;<?php echo $errors['formset-id']; ?></span>
            </td>
        </tr>
         <tr>
